+++ lib/l10n/app_localizations_en.dart
@@ -77,6 +77,28 @@ class AppL10nEn extends AppL10n {
   String get tooltipUnlike => 'Remove like';
 
   @override
+  String get recentlyPlayed => 'Recently played';
+
+  @override
+  String get tooltipSleepTimer => 'Sleep timer';
+
+  @override
+  String get sleepTimerOff => 'Turn off timer';
+
+  @override
+  String get playbackSection => 'Playback';
+
+  @override
+  String get crossfade => 'Crossfade';
+
+  @override
+  String get crossfadeSubtitle =>
+      'Fade out the end of a track and fade in the next';
+
+  @override
+  String get crossfadeDuration => 'Crossfade duration';
+
+  @override
   String songAdded(String title) {
     return '$title added';
   }