+++ test/screens/songs_screen_tablet_test.dart
@@ -12,6 +12,8 @@ import 'package:music_player_flutter/services/audio_handler.dart';
 import 'package:music_player_flutter/services/download_service.dart';
 import 'package:music_player_flutter/services/liked_songs_service.dart';
 import 'package:music_player_flutter/services/player_service.dart';
+import 'package:music_player_flutter/services/recently_played_service.dart';
+import 'package:music_player_flutter/services/search_history_service.dart';
 import 'package:music_player_flutter/screens/songs_screen.dart';
 
 class MockApiService extends Mock implements ApiService {}
@@ -66,6 +68,10 @@ void main() {
           ChangeNotifierProvider<DownloadService>.value(value: downloadService),
           ChangeNotifierProvider<LikedSongsService>.value(
               value: LikedSongsService()),
+          ChangeNotifierProvider<RecentlyPlayedService>.value(
+              value: RecentlyPlayedService()),
+          ChangeNotifierProvider<SearchHistoryService>.value(
+              value: SearchHistoryService()),
           Provider<ApiService>.value(value: mockApi),
           ChangeNotifierProvider<AuthService>(
             create: (ctx) => AuthService(ctx.read<ApiService>()),